backend/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
pdfplumber
python-docx
openai
numpy